fix incorrect fountain texture

+++ src/scene.h
@@ -49,7 +49,7 @@ typedef struct Scene {
     int8_t **texture_colours_used;
     int32_t **texture_colour_list;
     int8_t *texture_dimension;
-    int32_t *texture_loaded_number;
+    int64_t *texture_loaded_number;
     int32_t **texture_pixels;
     int8_t *texture_back_transparent;
     int32_t **texture_colours_64;
